pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hypertrack"
version = "0.1.0"
description = "A Python package for tracking Hyperliquid trading activity and detecting profitable traders"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = [
    "hyperliquid",
    "trading",
    "cryptocurrency",
    "defi",
    "blockchain",
    "trading-bot",
    "ml",
    "machine-learning",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial :: Investment",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.28.0",
    "websocket-client>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/hypertrack"
Documentation = "https://github.com/yourusername/hypertrack#readme"
Repository = "https://github.com/yourusername/hypertrack"
"Bug Tracker" = "https://github.com/yourusername/hypertrack/issues"

[project.scripts]
hypertrack = "hypertrack.recorder:main"
hypertrack-chain = "hypertrack.chain_subscriber:main"
hypertrack-profile = "hypertrack.wallet_profiler:main"

[tool.setuptools]
packages = ["hypertrack"]

[tool.setuptools.package-data]
hypertrack = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
